+++ src/Helper/RootAccessable.php
@@ -17,40 +17,18 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-namespace Art4\JsonApiClient;
+namespace Art4\JsonApiClient\Helper;
 
-use Art4\JsonApiClient\Input\Input;
+use Art4\JsonApiClient\Accessable;
+use Art4\JsonApiClient\Element;
+use Art4\JsonApiClient\Manager;
 
 /**
- * Manager Interface
+ * RootAccessable
+ *
+ * @internal
  */
-interface Manager
+class RootAccessable implements Accessable
 {
-    /**
-     * Parse the input
-     *
-     * @param Art4\JsonApiClient\Input\Input $input
-     *
-     * @throws Art4\JsonApiClient\Exception\ValidationException If $input contains invalid JSON API
-     *
-     * @return Art4\JsonApiClient\Accessable
-     */
-    public function parse(Input $input);
-
-    /**
-     * Get a factory from the manager
-     *
-     * @return Art4\JsonApiClient\Factory
-     */
-    public function getFactory();
-
-    /**
-     * Get a param by key
-     *
-     * @param string $key
-     * @param mixed  $default
-     *
-     * @return mixed
-     */
-    public function getParam($key, $default);
+    use AccessableTrait;
 }